backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
httpx==0.28.1
supabase==2.11.0
google-genai==1.14.0
yfinance==0.2.51
fredapi==0.5.2
apscheduler==3.11.0
python-dotenv==1.0.1
